requirements.txt
aiogram==3.0.0
telegram
python-telegram-bot
python-dotenv
